zerodivision correct

+++ ait/components/profile/benchmark/test/ST/test_infer_bert.py
@@ -206,12 +206,10 @@ class TestClass():
         assert math.fabs(msame_inference_time_ms) > TestCommonClass.EPSILON
         # compare
         allowable_performance_deviation = 0.01
-        try:
-            actual_performance_deviation = math.fabs(msame_inference_time_ms - \
-                                                     ais_bench_inference_time_ms)/msame_inference_time_ms
-        except ZeroDivisionError:
-            logger.error("zero division!")
-        assert actual_performance_deviation < allowable_performance_deviation
+        if msame_inference_time_ms == 0:
+            raise ZeroDivisionError
+        assert math.fabs(msame_inference_time_ms - ais_bench_inference_time_ms)/msame_inference_time_ms \
+            < allowable_performance_deviation
         os.remove(msame_infer_log_path)
         shutil.rmtree(output_dir_path)
